Add special case handling for 'force' option (#27)
The '--force' option is related to '--max-configs', but instead of
setting the maximum nuber of configs to check (which '--max-configs'
does), '--force' ensures that all configs are always checked, regardless
of how many they are. Since they are related in this fashion, instead of
adding a specific configuration option for '--force' it's added as a
special case of the max_configs setting. That way there should never be
any uncertainty as to what option or setting is actually applied in the
call to cppcheck itself.